backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
argon2-cffi>=21.0.0

# Database
mysqlclient>=2.2.0

# Vector Database for Face Embeddings
chromadb>=0.4.0

# Celery & Redis
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Image Processing & ML
numpy>=1.26.0
Pillow>=10.2.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
deepface>=0.0.79
tf-keras>=2.15.0
insightface>=0.7.3
onnxruntime>=1.16.0
mediapipe>=0.10.9

# Environment & Config
python-decouple>=3.8
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# API Documentation
drf-spectacular>=0.27.0

# Security
cryptography>=41.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
